Refactors Profile as component, adds scroll behavior 'auto' to resetView, cleans comments
